kernel: use prebuilt by defalult
to build the kernel inline specifiy BUILD_KERNEL=true on the
command line. since the source is not automatically synced we let it
fallback to prebuilt even if BUILD_KERNEL is defined when the source
dir is not present.
to sync the source run find_deps after setting environment with envsetup.sh

Change-Id: Ic80cdd1ba0934d96ec75a9f11db26944f536f6a0