Remove security date from updater script
This conflicts with the overriding of security patch date introduced in
the commit below. Instead of adding yet another prop to counter that
and then have to maintain that prop, is just easier to remove this one.
I know is 'lazy' or whatever but really, no point in introducting yet
another prop.

https://github.com/DirtyUnicorns/android_build/commit/2a532db6e1eeb8e6d12b2ba0242b10d4830f3d66

FYI we're overriding the security patch date on certain devices to help
pass CTS/safety net on devices were the OEM is taking their sweet time
updating vendor (OnePlus, looking at you).

Change-Id: Ic3b93cdb42d7eb0ae8a4cd68e44a4dffbcfa779f